Refactored structure. Added readme.
