+++ .idea/workspace.xml
@@ -4,7 +4,7 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="b8c9e775-d1e1-4255-9886-31c01930e6dd" name="Changes" comment="Second task is completed. Still have to add readme though.">
+    <list default="true" id="b8c9e775-d1e1-4255-9886-31c01930e6dd" name="Changes" comment="Adjusted package naming.">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -45,7 +45,8 @@
   &quot;keyToString&quot;: {
     &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
     &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
-    &quot;SHARE_PROJECT_CONFIGURATION_FILES&quot;: &quot;true&quot;
+    &quot;SHARE_PROJECT_CONFIGURATION_FILES&quot;: &quot;true&quot;,
+    &quot;last_opened_file_path&quot;: &quot;/Users/artemkot/java/study/projects/praktikum/sprint_5/Diplom_Artem_Kot_12/diplom-2&quot;
   }
 }</component>
   <component name="RecentsManager">
@@ -236,7 +237,14 @@
       <option name="project" value="LOCAL" />
       <updated>1671225848225</updated>
     </task>
-    <option name="localTasksCounter" value="13" />
+    <task id="LOCAL-00013" summary="Adjusted package naming.">
+      <created>1671226948363</created>
+      <option name="number" value="00013" />
+      <option name="presentableId" value="LOCAL-00013" />
+      <option name="project" value="LOCAL" />
+      <updated>1671226948363</updated>
+    </task>
+    <option name="localTasksCounter" value="14" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -262,6 +270,7 @@
     <MESSAGE value="Implemented UserApi. Added prototype test." />
     <MESSAGE value="Finished order creation test case. Refactored everything, so it would be more easy to inject user creation into order-related tests." />
     <MESSAGE value="Second task is completed. Still have to add readme though." />
-    <option name="LAST_COMMIT_MESSAGE" value="Second task is completed. Still have to add readme though." />
+    <MESSAGE value="Adjusted package naming." />
+    <option name="LAST_COMMIT_MESSAGE" value="Adjusted package naming." />
   </component>
 </project>
