+++ src/test/java/site/stellarburgers/GetOrdersTest.java
@@ -1,8 +1,8 @@
-package site.stellarburgers.user;
+package site.stellarburgers;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import site.stellarburgers.user.common.OrderSteps;
+import site.stellarburgers.common.OrderSteps;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.equalTo;
