+++ src/test/java/site/stellarburgers/LoginUserTest.java
@@ -1,8 +1,8 @@
-package site.stellarburgers.user;
+package site.stellarburgers;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
-import site.stellarburgers.user.common.UserSteps;
+import site.stellarburgers.common.UserSteps;
 import org.junit.Test;
 
 import static org.hamcrest.CoreMatchers.equalTo;
